+++ src/app/actions.py
@@ -1,5 +1,6 @@
 
 from typing import Optional
+from pyrogram.errors.exceptions.forbidden_403 import UserIsBlocked
 
 try:
     from src.utils import log
@@ -11,6 +12,7 @@ try:
         MangaChapter,
         search_chapter_in_list
     )
+    from src.domain.communications import Chat
     from src.domain.communications import Suscription
     from src.app.communications import notify_suscribers
     from src.app.client import pyrogram_client, memory
@@ -24,6 +26,7 @@ except ModuleNotFoundError:
         MangaChapter,
         search_chapter_in_list
     )
+    from domain.communications import Chat
     from domain.communications import Suscription
     from app.communications import notify_suscribers
     from app.client import pyrogram_client, memory
@@ -33,12 +36,18 @@ except ModuleNotFoundError:
 #                                  MESSAGES                                   #
 ###############################################################################
 
-async def process_reporting():
-    """Reviews the last chapters and notifies the suscribers"""
+async def process_reporting() -> list[tuple[Suscription, Exception]]:
+    """Reviews the last chapters and notifies the suscribers
+    
+    Returns the list of suscriptions paired with errors, aimed to detect
+    problems like user blocks, etc
+    """
     assert pyrogram_client is not None, "Pyrogram client is None"
 
     suscriptions: list[Suscription] = memory.read_suscriptions()
     mangas: list[Manga] = memory.read_mangas()
+
+    report: list[tuple[Suscription, Exception]] = []
 
     for sus in suscriptions:
         sus_manga: Optional[Manga] = search_manga_by_name(
@@ -63,19 +72,93 @@ async def process_reporting():
         if not sus_manga.last_chapter or \
            sus.last != sus_manga.last_chapter.name:
 
-            await notify_suscribers(sus_manga.last_chapter, [sus.chat])
+            notification_statuses: list[tuple[Chat, Exception]] = \
+                await notify_suscribers(sus_manga.last_chapter, [sus.chat])
+            
+            if notification_statuses[0][1] is None:
 
-            if not memory.update_suscription_last(
-                chat_id=sus.chat.id,
-                manga_name=sus.manga.name,
-                last_chapter=sus_manga.last_chapter.name
-            ):
-                log("bot", "warning", [
-                    "process_reporting",
-                    f"Last chapter updated for chat {sus.chat.id} and "
-                    f"manga {sus.manga.name}"
-                ])
+                if not memory.update_suscription_last(
+                    chat_id=sus.chat.id,
+                    manga_name=sus.manga.name,
+                    last_chapter=sus_manga.last_chapter.name
+                ):
+                    log("bot", "warning", [
+                        "process_reporting",
+                        f"Last chapter updated for chat {sus.chat.id} and "
+                        f"manga {sus.manga.name}"
+                    ])
 
+            report.append(
+                sus,
+                notification_statuses[0][1]
+            )
+
+    return report
+
+
+###############################################################################
+#                                 ON MEMORY                                   #
+###############################################################################
+
+def delete_suscription(sus: Suscription) -> bool:
+    """
+    Wraps the full suscription deletion based on the suscription object
+    """
+    if sus is None:
+        log(
+            "bot",
+            "info",
+            [
+                "delete_suscription",
+                "A suscription was not provided: None"
+            ]
+        )
+
+    else:
+        return memory.delete_suscription(sus.chat.id, sus.manga.name)
+    
+    return False
+
+
+def prune_suscriptions(suscriptions: list[tuple[Suscription, Exception]]):
+    """
+    Deletes a list of suscriptions based on their issues report
+    """
+
+    for sus, ex in suscriptions:
+
+        if ex is None:
+            continue
+
+        if isinstance(ex, UserIsBlocked):
+            ex: UserIsBlocked
+            log(
+                "bot",
+                "error",
+                [
+                    "prune_suscriptions",
+                    "Suscription obsolete due to User Blocked, pruning ongoing..."
+                ]
+            )
+            log(
+                "bot",
+                "error",
+                [
+                    "prune_suscriptions",
+                    f"{ex.MESSAGE}"
+                ]
+            )
+            delete_suscription(sus)
+
+        else:
+            log(
+                "bot",
+                "warning",
+                [
+                    "prune_suscriptions",
+                    f"Unmanaged error on suscription '{ex.__class__}': no action scheduled"
+                ]
+            )
 
 ###############################################################################
 #                                  SCRAPPING                                  #